Move ExternalIPsConfig to its own file.
