+++ collector/lib/CollectorConfig.h
@@ -14,6 +14,7 @@
 #include <internalapi/sensor/collector.pb.h>
 
 #include "CollectionMethod.h"
+#include "ExternalIPsConfig.h"
 #include "HostConfig.h"
 #include "Logging.h"
 #include "NetworkConnection.h"
@@ -93,32 +94,6 @@ class CollectorConfig {
   bool IsProcessesListeningOnPortsEnabled() const { return enable_processes_listening_on_ports_; }
   bool ImportUsers() const { return import_users_; }
   bool CollectConnectionStatus() const { return collect_connection_status_; }
-
-  // Encapsulates the configuration of the External-IPs feature
-  class ExternalIPsConfig {
-   public:
-    enum Direction {
-      NONE = 0,
-      INGRESS = 1 << 0,
-      EGRESS = 1 << 1,
-      BOTH = INGRESS | EGRESS,
-    };
-
-    // Are External-IPs enabled in the provided direction ?
-    bool IsEnabled(Direction direction) { return (direction & direction_enabled_) == direction; }
-
-    // Direction in which External-IPs are enabled
-    Direction GetDirection() const { return direction_enabled_; }
-
-    // Extract the External-IPs configuration from the provided runtime-conf.
-    // If the runtime-configuration is unset then 'default_enabled' is used
-    // as a fallback to enable in both directions.
-    // 'runtime_config' should be locked prior to calling.
-    ExternalIPsConfig(std::optional<sensor::CollectorConfig> runtime_config, bool default_enabled);
-
-   private:
-    Direction direction_enabled_;
-  };
 
   // EnableExternalIPs will check for the existence
   // of a runtime configuration, and defer to that value
@@ -287,6 +262,5 @@ class CollectorConfig {
 };
 
 std::ostream& operator<<(std::ostream& os, const CollectorConfig& c);
-std::ostream& operator<<(std::ostream& os, const collector::CollectorConfig::ExternalIPsConfig& config);
 
 }  // end namespace collector
